Move cli_parser to fractool cli code (specific to ftcli)
