+++ src/ftcore/CMakeLists.txt
@@ -1,6 +1,5 @@
 add_library(ftcore 
     config.cpp
-    cli_parser.cpp
     map_color.cpp
     write_image.cpp
     generate_mandelbrot.cpp
@@ -10,7 +9,6 @@ target_link_libraries(ftcore
         png_static
         Boost::log 
         Boost::log_setup
-        Boost::program_options
         Boost::thread
         Boost::filesystem)
 install(TARGETS ftcore LIBRARY)
